+++ jwst/cube_build/cube_build_step.py
@@ -46,14 +46,15 @@ SHORT,MEDIUM,LONG, or ALL
          offset_list = string(default='NA')  # A file for dithered data containing additional ra and dec offsets
          wavemin = float(default=None)  # Minimum wavelength to be used in the IFUCube
          wavemax = float(default=None)  # Maximum wavelength to be used in the IFUCube
+         single = boolean(default=false) # Internal pipeline option used by mrs_imatch and outlier detection
          xdebug = integer(default=None) # debug option, x spaxel value to report information on
          ydebug = integer(default=None) # debug option, y spaxel value to report information on
          zdebug = integer(default=None) # debug option, z spaxel value to report  information on
-         single = boolean(default=false) # Internal pipeline option used by mrs_imatch and outlier detection
          output_type = option('band','channel','grating','multi',default='band') # Type IFUcube to create. Options=band,channel,grating,multi
          search_output_file = boolean(default=false)
          output_use_model = boolean(default=true) # Use filenames in the output models
        """
+
     reference_file_types = ['cubepar', 'resol']
 
     def process(self, input):
@@ -129,7 +130,7 @@ SHORT,MEDIUM,LONG, or ALL
 
         if self.single:
             self.output_type = 'single'
-#________________________________________________________________________________
+______________________________________________________________
 # Set up a list for Channel, Band, Grating and Filter
 # These will be filled in either from
 # 1. If user has not set them  the are filled in by what is contained in the