+++ jwst/jump/__init__.py
@@ -1,3 +1,3 @@
 from .jump_step import JumpStep
 
-__version__ = '0.9.3'
+__all__ = ['JumpStep']