+++ jwst/lastframe/__init__.py
@@ -1,3 +1,3 @@
 from .lastframe_step import LastFrameStep
 
-__version__ = '0.9.3'
+__all__ = ['LastFrameStep']