+++ jwst/saturation/__init__.py
@@ -1,3 +1,3 @@
 from .saturation_step import SaturationStep
 
-__version__ = '0.9.3'
+__all__ = ['SaturationStep']